src/requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
fastapi==0.71.0
h11==0.12.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
sse-starlette==0.10.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
